Refactor: Create service for product business logic, refactor product controlelr
