--- server/routes/products.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const router = express.Router()
-
-const productsController = require('../controllers/products')
-
-router.get('/', productsController.getProducts)
-
-module.exports = router
