+++ server/routes/products.route.js
@@ -1,0 +1,8 @@
+const express = require('express')
+const router = express.Router()
+
+const { getProducts } = require('../controllers/product.controller')
+
+router.get('/', getProducts)
+
+module.exports = router
